+++ e2e/tests/services.rs
@@ -4,7 +4,7 @@ use contracts::{
 use ethcontract::{prelude::U256, H160};
 use model::DomainSeparator;
 use orderbook::{
-    account_balances::Web3BalanceFetcher, api::validation::OrderValidator, database::Postgres,
+    account_balances::Web3BalanceFetcher, api::order_validation::OrderValidator, database::Postgres,
     event_updater::EventUpdater, fee::EthAwareMinFeeCalculator, metrics::Metrics,
     orderbook::Orderbook, solvable_orders::SolvableOrdersCache,
 };
