+++ orderbook/src/main.rs
@@ -7,7 +7,7 @@ use model::{
 };
 use orderbook::{
     account_balances::Web3BalanceFetcher,
-    api::validation::OrderValidator,
+    api::order_validation::OrderValidator,
     database::{self, orders::OrderFilter, Postgres},
     event_updater::EventUpdater,
     fee::EthAwareMinFeeCalculator,
